SlidingTilePuzzle - Update commit - Hour 4
Updated the difficulty dropdown menu, so that it goes in front of the tiles instead of behind. Updated the animation to look a bit nicer. Updated the shuffle function. Before this, the shuffle function was just generating random tiles, and might end up with unsolvable conditions; now the shuffle function works by simulating 1000 moves to shuffle them