SlidingTilePuzzle - Update commit - Hour 5
my most productive session yet ;p    . Updated the code so that now I can upload my own 1:1 image using tkinter file selection. It will check whether the image is 1:1 ratio, and will reject images that are not 1:1.If the gamemode selected is 3x3, it will cut the image 9 pieces using pillow and discard the last one(the empty tile). If the gamemode selected is 4x4, it will cut the image into 15 pieces and discard the last one(empty tile). Updated animation to remove ghosting of previous tiles. and added a scorekeeping function based on time and moves